requirements.txt
Django>=4.2.0,<5.0.0
django-health-check
djangorestframework
django-dbconn-retry
django-debug-toolbar
setuptools
networkx
python-dateutil
shapely
gunicorn
pyproj
pyyaml
pika==0.13
raven
numpy
gevent
celery==4.*
coverage
pytest-cov
requests
ortools
osmnx
psycopg2-binary==2.8.6